refactor: Extract input styling so we can re-use it for the select component
